+++ centrifuge/__init__.py
@@ -1,1 +1,4 @@
+# coding: utf-8
+# Copyright (c) Alexandr Emelin. MIT license.
 
+__version__ = "0.5.9"